delete user.js
