+++ modules/library.js
@@ -1,7 +1,31 @@
+/* eslint-disable no-use-before-define */
+/* eslint-disable max-classes-per-file */
 /* eslint-disable no-unused-vars */
 /* eslint-disable import/prefer-default-export */
-import UI from './user.js';
 import Book from './constructor.js';
+
+export class UI {
+  static showBooks(collection) {
+    const table = document.querySelector('.table');
+    table.innerHTML = '';
+    collection.forEach((book, index) => {
+      const row = document.createElement('tr');
+      const td1 = document.createElement('p');
+      td1.innerHTML = `<strong>"${book.title}" by ${book.author}</strong>`;
+      const td2 = document.createElement('div');
+      const button = document.createElement('button');
+      button.className = 'remove';
+      button.innerHTML = 'Remove';
+      button.addEventListener('click', () => {
+        library.removeBook(book.id);
+      });
+      td2.appendChild(button);
+      row.append(td1, td2);
+      table.appendChild(row);
+    });
+    document.querySelector('.form').reset();
+  }
+}
 
 export class Library {
   constructor() {
